remove CRM_Lineitemedit_Form_Add form

+++ lineitemedit.php
@@ -188,7 +188,7 @@ function lineitemedit_civicrm_pre($op, $entity, $entityID, &$params) {
         }
         else {
           if ($taxEnabled) {
-            $lineItemParams[$i]['tax_amount'] = CRM_Utils_Array::value($i, $params['item_tax_amount'], 0.00);
+            $lineItemParams[$i]['tax_amount'] = (float) CRM_Utils_Array::value($i, $params['item_tax_amount'], 0.00);
             $params['tax_amount'] += $lineItemParams[$i]['tax_amount'];
           }
           $params['total_amount'] = $params['net_amount'] = $params['amount'] += (CRM_Utils_Array::value('line_total', $lineItemParams[$i], 0.00) + CRM_Utils_Array::value('tax_amount', $lineItemParams[$i], 0.00));
